Muda tests relacionados ao item SuaMesa para uma suite de teste dedicada
